Refactor: extract debugger functionality to own file
