+++ src/extension.ts
@@ -17,9 +17,9 @@
  */
 
 import {
+  window,
   workspace,
   commands,
-  debug,
   ExtensionContext,
   Disposable,
   TextDocument,
@@ -34,44 +34,19 @@ import {
   TaskPanelKind,
   TaskGroup,
   ProcessExecution,
-  window,
-  ProgressLocation,
-  DebugAdapterDescriptorFactory,
-  DebugConfigurationProvider,
-  CancellationToken,
-  DebugConfiguration,
-  DebugAdapterDescriptor,
-  DebugAdapterExecutable,
-  DebugSession,
-  ProviderResult,
-  OutputChannel,
+  ProgressLocation,  
 } from 'vscode';
+
+import { activateDebugger } from './debugger';
 
 import { languageId } from './constants';
 import Client from './client';
 import findNargo from './find-nargo';
-import findNearestPackageFrom from './find-nearest-package';
 import { lspClients, editorLineDecorationManager } from './noir';
-
-import { spawn } from 'child_process';
 
 const activeCommands: Map<string, Disposable> = new Map();
 
 let outputChannel: OutputChannel;
-
-class NoirDebugAdapterDescriptorFactory implements DebugAdapterDescriptorFactory {
-  async createDebugAdapterDescriptor(
-    _session: DebugSession,
-    _executable: DebugAdapterExecutable,
-  ): ProviderResult<DebugAdapterDescriptor> {
-    const config = workspace.getConfiguration('noir');
-
-    const configuredNargoPath = config.get<string | undefined>('nargoPath');
-    const nargoPath = configuredNargoPath || findNargo();
-
-    return new DebugAdapterExecutable(nargoPath, ['dap']);
-  }
-}
 
 const activeMutex: Set<string> = new Set();
 
@@ -352,8 +327,6 @@ async function didChangeWorkspaceFolders(event: WorkspaceFoldersChangeEvent) {
 }
 
 export async function activate(context: ExtensionContext): Promise<void> {
-  outputChannel = window.createOutputChannel('NoirDebugger');
-
   const didOpenTextDocument$ = workspace.onDidOpenTextDocument(didOpenTextDocument);
   const didChangeWorkspaceFolders$ = workspace.onDidChangeWorkspaceFolders(didChangeWorkspaceFolders);
   const restart$ = commands.registerCommand('noir.restart', restartAllClients);
@@ -365,96 +338,11 @@ export async function activate(context: ExtensionContext): Promise<void> {
     context.subscriptions.push(disposable);
   }
 
-  context.subscriptions.push(
-    debug.registerDebugAdapterDescriptorFactory('noir', new NoirDebugAdapterDescriptorFactory()),
-    debug.registerDebugConfigurationProvider('noir', new NoirDebugConfigurationProvider()),
-    debug.onDidTerminateDebugSession(() => {
-      outputChannel.appendLine("Debug session ended.");
-    }),
-  );
+  activateDebugger(context);
 }
 
 export async function deactivate(): Promise<void> {
   for (const client of lspClients.values()) {
     await client.stop();
-  }
-}
-
-/**
- * Takes stderr or stdout output from the Nargo's DAP
- * preflight check and formats it in an Output pane friendly way,
- * by removing all special characters.
- *
- * Note: VS Code's output panes only support plain text.
- *
- */
-function preflightCheckPrinter(buffer: Buffer, output: OutputChannel) {
-  const formattedOutput = buffer.toString()
-        .replace(/\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])/g, '')
-        .replace(/[^ -~\n\t]/g, '');
-
-  output.appendLine(formattedOutput);
-}
-
-class NoirDebugConfigurationProvider implements DebugConfigurationProvider {
-  async resolveDebugConfiguration(folder: WorkspaceFolder | undefined, config: DebugConfiguration, token?: CancellationToken): ProviderResult<DebugConfiguration> {
-    if (config.program || config.request == 'attach')
-      return config;
-
-    if (window.activeTextEditor?.document.languageId != 'noir')
-      return window.showInformationMessage("Select a Noir file to debug");
-
-    const currentFilePath = window.activeTextEditor.document.uri.fsPath;
-    let currentProjectFolder = findNearestPackageFrom(currentFilePath);
-
-    const workspaceConfig = workspace.getConfiguration('noir');
-    const nargoPath = workspaceConfig.get<string | undefined>('nargoPath') || findNargo();
-
-    outputChannel.clear();
-    outputChannel.appendLine(`Using nargo at ${nargoPath}`);
-    outputChannel.appendLine("Compiling Noir project...");
-    outputChannel.appendLine("");
-
-    // Run Nargo's DAP in "pre-flight mode", which test runs
-    // the DAP initialization code without actually starting the DAP server.
-    // This lets us gracefully handle errors that happen *before*
-    // the DAP loop is established, which otherwise are considered
-    // "out of band".
-    const preflightCheck = spawn(nargoPath, [
-      'dap',
-      '--preflight-check',
-      '--preflight-project-folder',
-      currentProjectFolder
-    ]);
-
-    // Create a promise to block until the preflight check child process
-    // ends.
-    let ready: (r: Boolean) => void;
-    const preflightCheckMonitor = new Promise((resolve) => ready = resolve);
-
-    preflightCheck.stderr.on('data', ev_buffer => preflightCheckPrinter(ev_buffer, outputChannel));
-    preflightCheck.stdout.on('data', ev_buffer => preflightCheckPrinter(ev_buffer, outputChannel));
-    preflightCheck.on('data', ev_buffer => preflightCheckPrinter(ev_buffer, outputChannel));
-    preflightCheck.on('exit', async code => {
-      if (code !== 0) {
-        outputChannel.appendLine(`Exited with code ${code}`);
-      }
-      ready(code == 0);
-    });
-
-    if (!await preflightCheckMonitor) {
-      outputChannel.show();
-      throw new Error("Error launching debugger. Please inspect the Output pane for more details.");
-    } else {
-      outputChannel.appendLine("Starting debugger session...");
-    }
-
-    return {
-      type: 'noir',
-      name: 'Noir binary package',
-      request: 'launch',
-      program: currentFilePath,
-      projectFolder: currentProjectFolder,
-    }
   }
 }
